requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
openai>=1.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
sse-starlette>=1.6.0
